* develop:
  Update to beta version
  update job detail screen
  Auto create calendar event when Cleaner applies to a job.
  Fix issue: out of bounds when swipe left more than half of the available options
  Fix issue #9
  Add default range when user registered successful. Update range on profile save.
  add validation into profileViewController
  fix postal code
  change Address Street to address 1. change Address unit to address 2
  add validate field into SignupFirstStepViewController

+++ CleanerSeeker/CleanerSeeker/Worker.swift
@@ -9,7 +9,7 @@
 import Parse
 
 class Worker: PFObject, PFSubclassing {
-    @NSManaged var searchRadius: Double
+    @NSManaged var searchRadius: Int
     @NSManaged var userRelationId: PFUser
     @NSManaged var location: PFGeoPoint
     // MARK: PFSubclassing Protocol methods
